Move SenML JSON class in their own package.
